PD#135579: fix for Android 7.0 64-bit
1.remove force-32bit-compilation
2.fix compile error
3.fix iconv from ucnv
4.add switch for disabling audio or video

+++ am_mw/am_closecaption/am_vbi/linux_vbi/linux_vbi.c
@@ -114,7 +114,7 @@ static AM_ErrorCode_t vbi_alloc_filter(AM_VBI_Device_t *dev, AM_VBI_Filter_t *fi
 	
 	dmx->fd[filter->id] = fd;
 	
-	filter->drv_data = (void*)fd;
+	filter->drv_data = (void*)(long)fd;
 	
 	return AM_SUCCESS;
 }
